chore: 프로젝트 초기화면 설정 및 SwiftLint 적용
