+++ ProjectManager/ProjectManager/Resource/Assets.xcassets/Contents.json
