+++ ProjectManager/ProjectManager/Resource/Info.plist
